backend/requirements.txt
python-docx
python-pptx
numpy
pandas
uuid
Flask
Flask-Cors
python-dotenv
langchain
langchain-huggingface
langchain-community
langchain-groq
langchain_core
torch
torchtext
faiss-cpu
pypdf
supabase
scikit-learn
huggingface-hub
sentence-transformers